Implement local variables
utilising the stack to store them for speed

currently using lambdas as block scopes for simplicity, will replace
with function implementation in the future

edge case: not built with the idea that you can have something on the
stack in front of a variable being defined, so something like (lambda (+
2 (def a 3))) executes incorrectly, by overwriting the value of 2 with
3, since it is in the stack slot that is normally associated with the
first variable. Electing not to fix this, since it will be resolved when
the lambda implementation is replaced by real functions.